:construction: UPDATE: withConnection
